requirements.txt
pennylane

jax
flax
optax
grain
dm-pix


datasets[jax]
transformers[flax]
evaluate

tqdm
pillow